requirements.txt
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
streamlit